Added new classes along with some new folders. Re organized some code. THis version has not been tested if the code builds or not
